Sort out reducer stuff, fix tsconfig
